refactor(update): extract pipeline ops into update-ops module
